chore(android, desktop): add hamburger menu
